Share one recipe match card across kitchen and shopping
The shopping picker's cards had drifted from the kitchen's match cards.
Replace both with a single RecipeMatchCard: thumbnail (chef-hat fallback),
kitchen-match progress computed against the shared stock, the missing
ingredients, and the add/remove list toggle. Kitchen recipe views gain
image fields so both tabs render identical cards; the picker keeps its
servings stepper as a card footer when selected.

From review: returning an item from the kitchen also clears any stale
tick (CodeRabbit), and parseStock returns fresh objects so a mutated
snapshot can't corrupt the shared empty-state constant (Greptile).
